Moved styles settings for offer card to const.js
